+++ packages/validator/index.ts
@@ -1,1 +1,2 @@
 export * from './src/email/index';
+export * from './src/mobile/index';